Fixed harmless warning

+++ src/cowpatty-4.6/cowpatty.c
@@ -993,7 +993,7 @@ int dictfile_attack(struct user_opt *opt, char *passphrase,
 		if (fret < 8 || fret > 63) {
 			if (opt->verbose) {
 				printf("Invalid passphrase length: %s (%u).\n",
-				       passphrase, strlen(passphrase));
+				       passphrase, (unsigned int) strlen(passphrase));
 			}
 			continue;
 		} else {
